+++ tests/test.php
@@ -3,7 +3,7 @@
 *  Contemplate
 *  Light-weight Templating Engine for PHP, Python, Node and client-side JavaScript
 *
-*  @version 0.6.1
+*  @version 0.6.2
 *  https://github.com/foo123/Contemplate
 *
 *  @inspired by : Simple JavaScript Templating, John Resig - http://ejohn.org/ - MIT Licensed
@@ -122,7 +122,7 @@ $main_template_data=array(
     //'sepright' => $sepright,
     'data_client' => json_encode($data),
     'render_server' => Contemplate::tpl('demo', $data),
-    'render_inline' => Contemplate::tpl('inlinetpl', $listdata, false, '<% %>')
+    'render_inline' => Contemplate::tpl('inlinetpl', $listdata, array('separators' => '<% %>'))
 );
 
 echo Contemplate::tpl('main', $main_template_data);