+++ tests/test.py
@@ -4,7 +4,7 @@
 #  Contemplate
 #  Light-weight Templating Engine for PHP, Python, Node and client-side JavaScript
 #
-#  @version 0.6.1
+#  @version 0.6.2
 #  https://github.com/foo123/Contemplate
 #
 #  @inspired by : Simple JavaScript Templating, John Resig - http://ejohn.org/ - MIT Licensed
@@ -137,7 +137,7 @@ main_template_data={
     #'sepright' : sepright,
     'data_client' : json.dumps(data),
     'render_server' : Contemplate.tpl('demo', data),
-    'render_inline' : Contemplate.tpl('inlinetpl', listdata, False, '<% %>')
+    'render_inline' : Contemplate.tpl('inlinetpl', listdata, {'separators':'<% %>'})
 }
 
 # a simple custom http server